First pass at a devcontainer
It's definitely missing some things, like getting all the tooling installed. And I'm not sure that what's here is correct, like the generated .cargo/config.toml and the Dockerfile. But it's a place to start.

Note that there are some known limitations with Zed's devcontainer support, such as that it doesn't support forwarding ports, and that it's only on the preview version of Zed.

The devcontainer was initially generated based on the rust+postgresql config with vscode, and then modified to make it closer to my preferred flavor of a devcontainer.

I also expect to need changes in order to be able to run this with worktrees for doing parallel work. It would be really nice to be able to clone the repo into a volume, but I haven't attempted to test that on any project, so I've delayed doing that here as well.